% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcpnn_mcmc_signal.R
\name{bcpnn_mcmc_signal}
\alias{bcpnn_mcmc_signal}
\title{BCPNN IC using MCMC for a \code{2x2} contingency table}
\usage{
bcpnn_mcmc_signal(a, b, c, d, alpha = 0.05, n_mcmc = 1e+05)
}
\arguments{
\item{a}{also referred to as \eqn{n_{11}} as this is the count of event of interest under exposure of interest}

\item{b}{also referred to as \eqn{n_{10}} as this is the count of \emph{not} event of interest under exposure of interest}

\item{c}{also referred to as \eqn{n_{01}} as this is the count of event of interest under \emph{not} exposure of interest}

\item{d}{also referred to as \eqn{n_{00}} as this is the count of \emph{not} event of interest under \emph{not} exposure of interest}
}
\description{
Bayesian Confidence Propagation Neural Network Information Critereon (BCPNN IC) using
Markov Chain Monte Carlo (MCMC) for a \code{2x2} contingency table
}
\details{
It is assumed that the contingency table under consideration has drugs/exposures in the rows and outcomes/events in the columns.

Let's assume we are interested in drug Y and outcome Z, the contingency table will look like this

\tabular{ccc}{
. \tab outcome Z \tab not outcome Z\cr
drug Y \tab \code{a}=\eqn{n_{11}} \tab \code{b}=\eqn{n_{10}} \cr
not drug Y \tab \code{c}=\eqn{n_{01}}  \tab \code{d}=\eqn{n_{00}} \cr
}

The function returns a \code{data.frame} with the following columns:

\itemize{
  \item \code{analysis}: names from the supplied vector \code{a} if present, the values 1:length(a) (with 0 padding) if not
  \item \code{n11}: drug Y and outcome Z count (\code{a})
  \item \code{drug margin}: margin for drug Y (\code{a + b})
  \item \code{event margin}: margin for outcome Z (\code{a + c})
  \item \code{n..}: total count from contingency table (\code{a + b + c + d})
  \item \code{E_n11}: expected value of \code{n11} based on margins (\code{(a + b) * (a + c) / (a + b + c + d)})
  \item \code{est_name}: \code{"ror"}
  \item \code{est_scale}: \code{"orig scale"}
  \item \code{est}: ROR estimate on original scale (not log_e)
  \item \code{var_scale}: \code{"ln"}
  \item \code{var_est}: variance of the log_e(ROR) estimate
  \item \code{sd_est}: standard deviation of the log_e(ROR) estimate
  \item \code{alpha}: alpha level supplied for the \code{100*(1-alpha)\%} confidence intervals
  \item \code{ci_lo}: lower bound of the \code{100*(1-alpha)\%} confidence interval
  \item \code{ci_hi}: upper bound of the \code{100*(1-alpha)\%} confidence interval
}
}
\examples{
# Singhal et al. p409 table 16. Int J Pharm Pharm Sci, Vol 7, Issue 6, 405-411
bcpnn_mcmc_signal(28, 942, 17, 31435)
bcpnn_mcmc_signal(c("Cisplatin-Ototoxicity" = 28), 942, 17, 31435)
bcpnn_mcmc_signal(122, 1320, 381, 31341)
bcpnn_mcmc_signal(c("Cisplatin-Ototoxicity" = 28, "Carboplatin-Pruritis" = 122), c(942, 1320), c(17, 381), c(31435, 31341))
}
\author{
Ty Stanford <tystan@gmail.com>
}
