% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_all_non_neg_scalar.R
\name{check_all_non_neg_scalar}
\alias{check_all_non_neg_scalar}
\title{Test whether input arguments are non-negative scalar}
\usage{
check_all_non_neg_scalar(..., warn_zeros = FALSE)
}
\arguments{
\item{...}{arbitrary number of inputs (each to be checked whether they are non-negative scalar)}

\item{warn_zeros}{\code{TRUE} will provide a warning if any 0 values, (\code{FALSE}, default) will ignore 0 counts}
}
\description{
Test whether input arguments are non-negative scalar
}
\details{
Five checks are made for each input:
\itemize{
  \item input must not be \code{NULL}
  \item input must not be or contain \code{NA} or infinite values
  \item input must be integer/numeric
  \item input must have a length of 1
  \item input must be 0 or greater in value
}

Each argument can therefore be either a numeric/integer scalar.

Returns TRUE if all passed arguments are non-negative scalar (pass the above tests).
An error is thrown if all passed arguments are NOT non-negative scalar
}
\examples{
# these return TRUE
check_all_non_neg_scalar(1, 2, 3e12)
check_all_non_neg_scalar(0, 2, 3e12)

# these will throw warnings and return TRUE
# check_all_non_neg_scalar(2, 0, warn_zeros = TRUE)

# these will throw errors
# check_all_non_neg_scalar(12.0000000000000000001, 1:2)
# check_all_non_neg_scalar(1, 2, 1:2)
# check_all_non_neg_scalar(1, character(0))
# check_all_non_neg_scalar(1, numeric(0))
# check_all_non_neg_scalar(-11, 2, 1:2)
# check_all_non_neg_scalar(-11, NA)
# check_all_non_neg_scalar(-11, c(1, NA))
# check_all_non_neg_scalar(-11.00001)
# check_all_non_neg_scalar(-11, NULL)
# check_all_non_neg_scalar(c(11.00001, 12.00000000001), 1:2)
# check_all_non_neg_scalar(11.00001, 1:2)
}
\author{
Ty Stanford <tystan@gmail.com>
}
