% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_lengths_equal.R
\name{check_lengths_equal}
\alias{check_lengths_equal}
\title{Test whether input arguments vectors of the same length}
\usage{
check_lengths_equal(...)
}
\arguments{
\item{...}{arbitrary number of vectors}
}
\description{
Test whether input arguments vectors of the same length
}
\details{
It is assumed the input has been checked using \code{check_all_positive_ints(...)} before using this function.
Each argument can therefore be either a numeric/integer scalar or vector.

Returns TRUE if all vectors supplied are of the same length. Error thrown if not.
}
\examples{
### these return TRUE
check_lengths_equal(1)
check_lengths_equal(1:3)
check_lengths_equal(1, 2, 3e12)
check_lengths_equal(1:3, 2:4, 3:5)
check_lengths_equal(12.0000000000000000000001, 2)

### these will throw errors
# check_lengths_equal(2, 1:2, 0)
# check_lengths_equal(1:9, 0:9)
}
\author{
Ty Stanford <tystan@gmail.com>
}
