% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_mult_int_overflow.R
\name{is_mult_int_overflow}
\alias{is_mult_int_overflow}
\title{Test whether multiplication of two integers will cause machine overflow}
\usage{
is_mult_int_overflow(int1, int2)
}
\arguments{
\item{int1}{the first integer value (may be a numeric/double)}

\item{int2}{the second integer value (may be a numeric/double)}
}
\description{
Test whether multiplication of two integers will cause machine overflow (if stored as an integer).
}
\details{
Two numbers are supplied and the result is boolean/logical answering the question,
"if we multiply the two numbers supplied, can the result be stored as an integer?"
That is, the supplied values need to be numeric or integer, but the function is not picky about type
beyond being able to multiple/divide values

The logic used is as follows:

if \code{a * b > max_int} where \code{a,b} integers ==> overflow, then....

\code{a > max_int / b} ==> overflow.

The maximum integer value (max_int above) that can be stored in your R session can be found using \code{.Machine$integer.max}.
}
\author{
Ty Stanford <tystan@gmail.com>
}
