% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nx2_to_signal.R
\name{nx2_to_signal}
\alias{nx2_to_signal}
\title{ROR, PRR, BCPNN (MCMC and norm) of a \code{nx2} contingency table}
\usage{
nx2_to_signal(
  nx2_tab,
  orig_scale = FALSE,
  alpha = 0.05,
  n_mcmc = 1e+05,
  min_events = 3,
  warn = FALSE
)
}
\arguments{
\item{nx2_tab}{A tabulation of counts with \code{n} >= rows and \code{2} columns. The first column is the event of interest, the second
being other events. The rows are \code{n} drugs/exposures.}

\item{orig_scale}{\code{FALSE} (default), do you want the estimates to be on the ratio scale (e.g., BCPNN is on the log2 scale)}

\item{alpha}{for construction of the \code{100*(1-alpha)\%} confidence intervals}

\item{n_mcmc}{number of MCMC simulations per \code{(a,b,c,d)}-tuple}

\item{min_events}{the minimum number of events of interest for each drug/exposure to generate signal calculations. Default is 3.}

\item{warn}{if the minimum number of events of interest not met, should a warning/note be printed to the console. Default is \code{FALSE}.}
}
\value{
The function returns a \code{data.frame} that is returned for \code{battery_signal()} for each of 
the \code{n} drugs/exposures.
}
\description{
ROR, PRR, BCPNN (MCMC and norm) of a \code{nx2} contingency table
}
\details{
In addition to \code{nx2_tab} representing counts with n >= rows and 2 columns, the \code{class} of \code{nx2_tab}
must be \code{"table"}. This is achieved by using \code{table()} on columns of record level data or converting a 
\code{matrix}/\code{data.frame} of counts using \code{as.table()}.
}
\examples{
set.seed(1234)
library(dplyr)
# some made up record level data
drg <- base::sample(paste0("drg_", LETTERS[1:5]), size = 1000, replace = TRUE)
outc <- base::sample(paste0("outc_", LETTERS[25:26]), size = 1000, replace = TRUE)
(drg_tab <- table(drg, outc))
nx2_to_signal(drg_tab, orig_scale = TRUE)
}
\author{
Ty Stanford <tystan@gmail.com>
}
