% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsexp_shrink_signal.R
\name{obsexp_shrink_signal}
\alias{obsexp_shrink_signal}
\title{Observed-to-expected ratios with shrinkage for a \code{2x2} contingency table}
\usage{
obsexp_shrink_signal(a, b, c, d, alpha1 = 0.5, alpha2 = 0.5, alpha = 0.05)
}
\arguments{
\item{a}{also referred to as \eqn{n_{11}}{n11} as this is the count of event of interest under exposure of interest}

\item{b}{also referred to as \eqn{n_{10}}{n10} as this is the count of \emph{not} event of interest under exposure of interest}

\item{c}{also referred to as \eqn{n_{01}}{n01} as this is the count of event of interest under \emph{not} exposure of interest}

\item{d}{also referred to as \eqn{n_{00}}{n00} as this is the count of \emph{not} event of interest under \emph{not} exposure of interest}

\item{alpha1}{(default \code{0.5}) the numerator shrinkage parameter \code{>=0}. See details.}

\item{alpha2}{(default \code{0.5}) the denominator shrinkage parameter \code{>=0}. See details.}

\item{alpha}{for construction of the \code{100*(1-alpha)\%} confidence intervals}
}
\value{
The function returns a \code{data.frame} with the following columns:

\itemize{
  \item \code{analysis}: names from the supplied vector \code{a} if present, the values 1:length(a) (with 0 padding) if not
  \item \code{n11}: drug Y and outcome Z count (\code{a})
  \item \code{drug margin}: margin for drug Y (\code{a + b})
  \item \code{event margin}: margin for outcome Z (\code{a + c})
  \item \code{n..}: total count from contingency table (\code{a + b + c + d})
  \item \code{E_n11}: expected value of \code{n11} based on margins (\code{(a + b) * (a + c) / (a + b + c + d)})
  \item \code{est_name}: \code{"obs exp ratio"}
  \item \code{est_scale}: \code{"log2"}
  \item \code{est}: OE ratio (with shrinkage) estimate on log2 scale 
  \item \code{var_scale}: \code{NA}
  \item \code{var_est}: \code{NA}
  \item \code{sd_est}: \code{NA}
  \item \code{alpha}: alpha level supplied for the \code{100*(1-alpha)\%} confidence intervals
  \item \code{ci_lo}: lower bound of the \code{100*(1-alpha)\%} confidence interval
  \item \code{ci_hi}: upper bound of the \code{100*(1-alpha)\%} confidence interval
}
}
\description{
Observed-to-expected ratios with imple shrinkag of Norén (2013) for a \code{2x2} contingency table
}
\details{
The observed to expected (OE) ratio with approximate \code{100*(1-alpha)\%} confidence intervals
are constructed on the log2 scale as outlined in Norén et al. (2013).

The observed to expected (OE) ratio with shrinkage estimates is simply \code{(O + alpha1) / (E + alpha2)}.

If \code{(O + alpha1)} < \code{1}, then the exact uncertainty limits should be used. 
That is the \code{100*(1-alpha)\%} confidence intervals as implemented in \code{bcpnn_mcmc_signal()} (Norén et al., 2013).

\code{log2(OE)} approximates the Bayesian confidence propagation neural network information component (IC) 
with reasonable accuracy when \code{alpha1 = alpha2 = 0.5} (Norén et al., 2013).

Note that the \code{a},\code{b},\code{c},\code{d} inputs may be vectors of equal length, for which the function
will perform the calculations for each individual \code{(a,b,c,d)}-tuple moving across the vectors. 

It is assumed that the contingency table under consideration has drugs/exposures in the rows and outcomes/events in the columns.

Let's assume we are interested in drug Y and outcome Z, the contingency table will look like this

\tabular{ccc}{
. \tab outcome Z \tab not outcome Z\cr
drug Y \tab \code{a}=\eqn{n_{11}}{n11} \tab \code{b}=\eqn{n_{10}}{n10} \cr
not drug Y \tab \code{c}=\eqn{n_{01}}{n01}  \tab \code{d}=\eqn{n_{00}}{n00} \cr
}
}
\examples{
# Singhal et al. p409 table 16. Int J Pharm Pharm Sci, Vol 7, Issue 6, 405-411
obsexp_shrink_signal(28, 942, 17, 31435)
obsexp_shrink_signal(
  c("Cisplatin-Ototoxicity" = 28), 942, 17, 31435
)
obsexp_shrink_signal(122, 1320, 381, 31341)
obsexp_shrink_signal(
  c("Cisplatin-Ototoxicity" = 28, "Carboplatin-Pruritis" = 122), 
  c(942, 1320), 
  c(17, 381), 
  c(31435, 31341)
)
# case when (O + alpha1) < 1 [see details]
set.seed(1234)
obsexp_shrink_signal(0, 12, 49, 300) # CI the same as below for bcpnn_mcmc
set.seed(1234)
bcpnn_mcmc_signal(0, 12, 49, 300) 

}
\references{
Norén GN, Hopstadius J, Bate A. 
Shrinkage observed-to-expected ratios for robust and transparent large-scale pattern discovery. 
Statistical methods in medical research. 2013 Feb;22(1):57-69.
}
\seealso{
\code{\link{bcpnn_mcmc_signal}}, 
\code{\link{bcpnn_norm_signal}}, 
\code{\link{ror_signal}}, 
\code{\link{prr_signal}}
}
\author{
Ty Stanford <tystan@gmail.com>
}
